requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
bidict==0.22.1
certifi==2019.11.28
chainlit==0.7.0
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
dataclasses-json==0.5.14
dbus-python==1.2.16
Deprecated==1.2.14
diskcache==5.6.3
distro-info==0.23+ubuntu1.1
exceptiongroup==1.1.3
fastapi==0.99.1
fastapi-socketio==0.0.10
filelock==3.12.4
filetype==1.2.0
frozenlist==1.4.0
fsspec==2023.9.2
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpcio==1.58.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.17.3
idna==2.8
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.304
langsmith==0.0.41
Lazify==0.4.0
llama-index==0.8.33
llama_cpp_python==0.2.7
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
nltk==3.8.1
nodeenv==1.8.0
numexpr==2.8.7
numpy==1.26.0
openai==0.28.0
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-instrumentation==0.40b0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
packaging==23.1
pandas==2.1.1
prisma==0.10.0
protobuf==4.24.3
pydantic==1.10.13
PyGObject==3.36.0
PyJWT==2.8.0
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.7.1
python-graphql-client==0.4.3
python-multipart==0.0.6
python-socketio==5.9.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-unixsocket==0.2.0
safetensors==0.3.3
simple-websocket==0.10.1
six==1.14.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
starlette==0.27.0
syncer==2.0.3
tenacity==8.2.3
tiktoken==0.5.1
tokenizers==0.13.3
tomli==2.0.1
tomlkit==0.12.1
tqdm==4.66.1
transformers==4.33.2
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
unattended-upgrades==0.1
uptrace==1.20.0
urllib3==1.25.8
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.2
zipp==3.17.0
